requirements.txt
# Flask
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7

# Compatible Embeddings + Transformers + HF Hub
sentence-transformers==2.2.2
transformers==4.29.2
huggingface-hub==0.14.1

# PyTorch
torch==2.2.0

# Vector database
faiss-cpu

# PDF parsing
pdfplumber==0.10.3

# Data
pandas==2.1.4
numpy==1.26.2

# Utilities
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.0
